requirements.txt
Flask
ffmpeg-python
openai-whisper
deep-translator
sumy